Migrate dialogs to Base UI and unmount them on close (#892)
* Migrate dialogs to Base UI and unmount them on close


* Split dialog form state into per-field useState hooks


* Use urlValid directly instead of aliasing it


---------
